Stats Project from Sophomore Year
This stats project was done in order to determine pitch speeds for starting and relief pitchers and determine whether any sort of conclusion could be drawn from the data. It was determined, using two baseball games from the start of the Pittsburgh Pirates 2019 season, that relief pitchers threw 2mph more than starting pitchers.